JPA: Delete Beer By Id Not Found

+++ src/main/java/mak/springframework/spring6restmvc/service/BeerServiceJPA.java
@@ -82,9 +82,12 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deleteById(UUID id) {
-
-        beerRepository.deleteById(id);
+    public Boolean deleteById(UUID id) {
+        if (beerRepository.existsById(id)) {
+            beerRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
